chore(server): split user module into more logical units
