+++ server/internal/core/user/user.go
@@ -1,39 +1,8 @@
 package user
 
 import (
-	"context"
-
-	"github.com/google/uuid"
-
 	"github.com/oxidrive/oxidrive/server/internal/core/password"
 )
-
-type ID (uuid.UUID)
-
-func EmptyID() ID {
-	return ID(uuid.UUID{})
-}
-
-func NewID() ID {
-	return ID(uuid.Must(uuid.NewV7()))
-}
-
-func ParseID(s string) (ID, error) {
-	id, err := uuid.Parse(s)
-	if err != nil {
-		return ID{}, err
-	}
-
-	return ID(id), nil
-}
-
-func (i ID) AsUUID() uuid.UUID {
-	return uuid.UUID(i)
-}
-
-func (i ID) String() string {
-	return i.AsUUID().String()
-}
 
 type User struct {
 	ID           ID
@@ -56,11 +25,4 @@ func Create(username string, pwd string) (*User, error) {
 
 func (u User) VerifyPassword(password string) (bool, error) {
 	return u.PasswordHash.Verify(password)
-}
-
-type Users interface {
-	Count(context.Context) (int, error)
-	Save(context.Context, User) (*User, error)
-	ByID(context.Context, ID) (*User, error)
-	ByUsername(context.Context, string) (*User, error)
 }
