+++ server/internal/core/user/users.go
@@ -1,0 +1,11 @@
+package user
+
+import "context"
+
+type Users interface {
+	Count(context.Context) (int, error)
+	Save(context.Context, User) (*User, error)
+	ByID(context.Context, ID) (*User, error)
+	ByUsername(context.Context, string) (*User, error)
+}
+
